fix(translations): add translations to header_plain.blade.php
Some Vue components were not picking up the translation groups from the PHP
files. This is because we were not setting the window.translations object
in the header_plain.blade.php file which is used by the login page. Other
pages were using the header.blade.php file which does set the object.