refactor(docker): update Dockerfile and docker_run.sh for improved dependency management
- Changed the installation method for Composer in Dockerfile to use the official installer.
- Updated docker_run.sh to use the `composer install` command directly instead of referencing the phar file.
- Removed unnecessary SSH service start command from docker_run.sh.
- Removed using `root` user for docker-compose.yml.

This slims and refactors the Dockerfile. We can utilize the docker php
extension installer to install the PHP extensions but also the system
dependencies that are required by the PHP extensions. As such, we no longer
need to install all the lib* system packages.

Additionally by using the official PHP image we no longer need to install
some of the PHP extensions as they are already installed in the image.

Lastly, we will no longer be using ssh to connect to the container.