checkpoint: got testing env file working
Instead of setting the env vars in the phpunit.xml file, I've created a new
.env.testing file that will be used to set the environment variables for the
tests.

This will help with reusing the same MySQL container for development and testing.

That said, I needed to manually customize the env loading since the .env.testing
file completely overrides the .env file or the .env file was not loaded.

As such, all the other env vars would be lost. Now we will load the
.env.testing file first and then the .env file. This way, the .env.testing
file will override the .env file.