checkpoint: got test db working
There were some issues with the previous approach. The test bootstrap file
was only affective when running the artisan test command, however we
also needed to run migrations. As such, the bootstrap file had no effect
and we ran into the same issue as before. The solution is to move the
environment bootstrap into the main artisan file so all commands are
affected.

In addition to this, we need to actually create the test database in the
docker container so an additional task was added to the Taskfile that
utilises the already sourced dotenv files in the global file while
overriding the test db env variables by sourcing the .env.testing file
within the respective task. After which we will also run the migrations
for the test database.