feat(env): add local environment files
We will need to revisit this in the future to clean up the usage of
environment files for different applications.

Laravel has a local environment which means it will pick up the .env.local
file and use it over the .env file. However, we won't be using the
.env.local file with that approach. Intsead it will be used in the Taskfile
to generate the final .env file from the .env.template file.

Laravel might still pickup the .env.local file if APP_ENV is set to local,
but we already added a workaround with environment.php to merge in the
.env file, as opposed to just using the .env.local file.