feat(translations): implement instance-aware translation loader
Added a custom TranslationServiceProvider to support instance-specific translations.
This will extend Laravel's lang structure to have separate translations
for each Restarters instance. As such, this allows us to retain the
translations functionality but allow content to be overridden for each instance.

While doing so I merged the RobustTranslator code into the TranslationServiceProvider,
to remove the file and contain all the logic within the service provider.

This also updates the .env files to include APP_INSTANCE variable for
dynamic instance configuration.