refactor(vue-lang): create custom lang.js
We have a file named lang.js but it uses an actual package called
lang.js that has not been updated for a couple of years now.
https://github.com/rmariuzzo/Lang.js

On top of that, the inner workings of the package no longer work with our new
lang folder structure that is subgrouped by instances. The generated
translations.js file has a structure of

instances.<instance>.<locale>.<group|key> : string|object

The lang.js package splits the key by the dots and assumes each part
is a new level of a nested object. As such, it expects 'instances' to
be a key pointing to an object while it should really be
instances.instance.locale.group|key. as the first level of the lookup.

To resolve this, we will create a custom lang.js file, that builds on the
original lang.js package, and allows us to lookup translations for
our given structure.

To also avoid confusion, the lang.js file will house the LangJS class
and a new file i18n.js will be created to house the original __() translation
functionality.