feat(debug): add appDebug variable for translation debugging
Let's make use of the debug option for lang.js to log missing translations.
We can utilize APP_DEBUG to enable this feature.

That said, there was an issue with the boolean value being returned from
the PHP env() function. It returns 1 for true, 0 for false instead of
true/false.