feat(errors): add error reporting translations for instances
Created new language files for error reporting in the base and Fixit Clinic instances.
This will allow us to customize the error reporting for each instance.
To actually render the HTML in the error translations we need to use the
{!! !!} syntax to not escape the HTML.

Additionally, it seems there was a bug that was introduced with our
Laravel Shift upgrade that introduced raw text
"use App\Providers\AppServiceProvider;" when in reality the Laravel
Shift upgrade was trying to utilize the AppServiceProvider.php file.
However, these files do not need to use the AppServiceProvider.php file
so we can remove it and therefore remove the raw text from being rendered
on the error pages.