fix(docker): run npm production after translation generation
The public/js/translations.js file was not correct because we were
copying the translations.js file from resources/js/translations.js before
we ran the lang:js artisan command. As such, the translations.js file
was out of sync. We need to run npm production after the lang:js
artisan command to ensure the translations.js file is correct.