fix(logo): update logo inclusion logic
If the APP_INSTANCE is set to `base`, then the restarters logo won't be
used since there is no logo-base.blade.php file. I chose not to add or
rename the restarters logo file to logo-base.blade.php since its not
really a base logo. Instead, I chose to update the conditional logic
to use restarters in both scenarios. Meaning, if the APP_INSTANCE is
set to base, then the restarters logo will be used, as well if APP_INSTANCE
is not set at all.