feat(user): enhance user creation command with role and consent options
Updated the user:create command to include options for role assignment and
consent dates. The docker_run.sh script has been modified to utilize the
new command format for creating an admin user with auto-consent.

This will also be useful for creating admin users during the initialization
of a new instance of the app.