feat(groups): update logic for group postcode field
Let's allow anyone to edit the postcode field. This means it won't
be disabled when creating a new group and when editing an existing
group while not being an admin.

I thought about just not rendering the postcode field at all but I did
not see the same restrictions to the other fields. As such, it made
more sense to me to allow anyone to edit the postcode field like they
could the other fields.

As part of this change, I updated the hint text to make it clear that
the postcode field will take precedence over the location's postcode. In
addition, when first creating a group, the postcode field will be
pre-filled with the location's postcode.