feat(groups): add override options for postcode and timezone
Enhanced the Group model and related components to support 'override_postcode'
and 'override_timezone' features. This will give us more control over the
group creation page as the current logic is somewhat confusing.

For example, if the fields are empty then the location's postcode and timezone
will be used. Afterwhich, if the location is changed, the postcode and timezone
will not be updated.

While we could update the fields every time the location is changed, this
would regress the ability to override the postcode and timezone.

As such, let's implement the "override" functionality in a more explicit
way by adding a checkbox to state whether the postcode and timezone should
be overridden.