fix(lang.js): add choice and getLocale methods
When creating a custom version of the lang.js package, I missed that
we used the choice and getLocale methods in a couple of places. As
such, things like the all groups page was broken because the choice
method was not available.

This adds the choice and getLocale methods to our custom lang.js file.