fix(image-uploads): use global var to retrieve image upload enabled for Vue components
If we build the Vue components then we can't utilize the
PHP if statements in Vue components. In dev we can because we utilize
a dev server.

As such, we need to do something similar like with the window.appInstance
and window.appDebug to retrieve the image upload enabled variable in
direct components. Otherwise, we'd need to propogate the variable, as a
prop, across multiple components.

Following the 3's rule, we'll create a parent global variable named
window.Laravel that will, in the future, hold all the Laravel specific
variables that we need to access in the Vue components. For now, we will
just store the image upload enabled variable and in another pull request
we'll add the other variables.