refactor(skills): catch case where we want to update skills description
In the case where we add a description to a skill in our JSON file, we
would not be able to update the record in the database. This is because
we only check if the skill exists in the database.

Let's add an additional check to see if the description is empty and we
want to update the description. In this case, we'll update the record.