fix(sentry): ensure Sentry DSN defaults to null if not set
Updated the Sentry configuration to default the DSN to null when neither
SENTRY_LARAVEL_DSN nor SENTRY_DSN are not set or are empty. These envs
are declared in the .env file but are empty. However, the Sentry Laravel
package expects a null value if we want to disable Sentry.