feat(taskfile): add dependency management tasks
Add tasks for clearing and installing Composer and NPM dependencies.
This also includes logic to wrap the tasks in execution context.
Essentially this logic simplifies the workflow by removing the need to run:

task docker:run:bash -- "task <task> -- [args]"

to just run:
task <task> -- [args]

To do this we need to pass the DOCKER_NAMESPACE variable to deps.yml
because included tasks don't inherit context from other included files.
Only the root Taskfile merges all tasks and provides complete context.
If the tasks in the deps file need to run the run:bash command from the
docker file, then we need to use the alias for how the docker files was
included in the root Taskfile.yml file. In this case the alias is "docker:".
If we ever change the alias we can just change the value of the DOCKER_NAMESPACE
variable instead of having to change all references in the deps.yml file.