feat(docker): update production configuration and entrypoint script
By utilizing the taskfiles we can tremendously simplify the production
entrypoint script. In addition, we will no longer be mounting or adding
a default .env file to the production image. Instead we will be sourcing
the .env.prod file to the runtime environment.

The only thing we will need to ensure is that the APP_KEY env variable
is set. Otherwise the app should not start. This is different than the
dev environment where we generate a key if it is not set.