fix(docker): update production nginx configuration and HTTPS middleware
This configures the production Nginx service to use a new nginx.conf file
and mounts the public directory to serve the static files.

In addition, we needed to update the HTTPS middleware. If we attempt to
use the production image locally then we will get SSL errors because
the app is configured to use HTTPS. As such, we will add a conditional
check to see if both APP_ENV is production and APP_DEBUG is true. This
scenario should only happen when we are testing the production image
locally. If this condition is met then we will downgrade the request
to use HTTP.