feat(docker): add health check functionality to production setup
Implemented a health check feature for the production environment. This includes
a new health check command, a health check script, and updates to the
docker-compose configuration to ensure service health is monitored.

There were some issues using the php-fpm-healthcheck script as the
php-fpm base image could not find any of the debian fcgi packages. Instead
of attempting to add a new repository to be able to install a fcgi package,
we can just use a custom healthcheck script and command to check
the health of the application. This should be less bloat on the image
size.