fix(helm): enhance security context configuration
Updated security context settings in deployment to ensure proper permissions
and added fsGroup support in values.yaml for volume permissions. This
was causing issues with the state of the container. ArgoCD continued to
state the container was out-of-sync because it was unable to apply the
fsGroup security context. Let's update our logic for file permissions
to ensure that the container is in a consistent state.