refactor(docker): simplify entrypoint script and remove .env creation
Removed the automatic creation of the .env file in the entrypoint script and
ensured that the script checks for the existence of the .env file and the
APP_KEY variable before proceeding.

As part of this change, I replaced the bash steps with the equivalent
Taskfile tasks.

Lastly, we don't need the wait for db step anymore because we have a
healthcheck in the docker-compose.prod.yml file.