feat(helm): share public files between PHP-FPM and Nginx containers
Again, similar to the production docker compose, we need to share the
public files between the PHP-FPM and Nginx containers. However, this time
around we will be mounting an empty directory so we need to copy the files
from the production image first. We will add an init container to do this.