feat(helm): add external secrets configuration for sensitive credentials
This will allow us to use External Secrets Operator without a direct
dependency on the External Secrets Operator in this Helm Chart.
We can disable the creation of the secrets in the values.yaml file and
use the External Secrets Operator to create said secrets.

The map keys are used for the map functionality in the application.

The database credentials are used to create to a database other than
the internal MySQL database.