fix(docker): set proper permissions for the shared storage volume
The nginx sidecar does not run as the www-data user and some of the
files in the public directory do not have group read permissions.

To address this we will ensure the docker image sets the proper permissions
both in the image and in the entrypoint script. In addition, we will default
to using the fsGroup of 33 (www-data) to set the proper permissions for
the shared storage volume.