feat(mail): add logo configurations for email templates
Added new logo configurations in mail.php for 'fixitclinic' and 'base' logos.
Updated the email message template to dynamically use the logo based on the
app instance.

That said we could not reuse the logo blade file that contains the svg
logo as SVGs are not widely supported in email clients. Instead we will
use a jpg image for the logo.