feat(auth): create user from iFixit session data
- Updated User model to include external_user_id and external_username fields.
- Implemented syncFromExternal method to create or update users from iFixit data.
- Enhanced UserController's logout method to handle redirection for external users.
- Added getCurrentUser method in iFixitAuthService to retrieve user data from iFixit session.
- Improved iFixitAuthStrategy to authenticate users and log them in based on iFixit session data.

This mainly adds the ability to create/update users from iFixit session data.
This will also set the user's role to Administrator if they are an Admin
in iFixit. That said, it will only affect the user's role on first creation.
If we manually change the user's role, then it will not be overwritten by
their iFixit role.