feat(auth): conditionalize frontend components
We still need to conditionalize some of the frontend components even after
addressing most of it with the routing.

These components revolve around changing a user's password and inviting
users to the platform.

We can't change the password for an externally created user so we need to
hide components related to that. In addition, we can't invite users to the
platform as they would need to create an account on iFixit first and
there is no current way to pre-register them.