feat: add migration tool for external auth users
If a user created via local auth has the same email address as an iFixit
user, then they will be unable to login to the site with their iFixit
account. We can't really delete the users from the DB if they have
already created groups and events. As such, this migration tool will
allow us to manually sync/migrate the users with their iFixit counterparts.